Refactor: 버튼 폴더명 변경
